+++ libnavigation-core/src/main/java/com/mapbox/navigation/core/reroute/LegacyRerouteControllerAdapter.kt
@@ -1,13 +1,15 @@
 package com.mapbox.navigation.core.reroute
 
+import com.mapbox.navigation.base.route.RouterOrigin
 import com.mapbox.navigation.base.route.toNavigationRoutes
 
 internal class LegacyRerouteControllerAdapter(
     private val legacyRerouteController: RerouteController
-) : NavigationRerouteController, RerouteController by legacyRerouteController {
+) : NavigationRerouteController,
+    RerouteController by legacyRerouteController {
     override fun reroute(callback: NavigationRerouteController.RoutesCallback) {
         legacyRerouteController.reroute { routes ->
-            callback.onNewRoutes(routes.toNavigationRoutes())
+            callback.onNewRoutes(routes.toNavigationRoutes(), RouterOrigin.Custom())
         }
     }
 }